Modificato i miei metodi permettendo l'inserimento dei parametri da utente + tradotto in inglese, spostato metodo productsAlreadyInWarehouse nella classe warehouse, aggiunto .toUpperCase quando si inserisce il tipo di prodotto
