+++ carello/src/progetto/Main.java
@@ -6,7 +6,7 @@ class Main {
     public static void main(String[] args) throws Exception {
 
         printWelcomeMsg();
-        productsAlreadyInWarehouse();
+        Warehouse.productsAlreadyInWarehouse();
         printMenuOperations();
         menuOperations();
         printGoodbyeMsg();
@@ -76,16 +76,13 @@ class Main {
                     System.out.println(Warehouse.findSellingPrice());
                     break;
                 case 8:
-                    System.out.println("Cerca prodotto con prezzo di acquisto: 80");
-                    System.out.println(Warehouse.findBuyingPrice(80));
+                    System.out.println(Warehouse.findBuyingPrice());
                     break;
                 case 9:
-                    System.out.println("Cerca prodotto con range di prezzo di vendita tra 50 e 200");
-                    System.out.println(Warehouse.findRangePrice(50, 200));
+                    System.out.println(Warehouse.findRangePrice());
                     break;
                 case 10:
-                    System.out.println("Trova prezzo medio di acquisto degli articoli TABLET");
-                    System.out.println(Warehouse.findAvgPrice("TABLET"));
+                    System.out.println(Warehouse.findAvgPrice());
                     break;
                 default:
                     System.out.println("Something went wrong");
@@ -94,40 +91,4 @@ class Main {
         } while (continueMenu);
     }
 
-    public static void productsAlreadyInWarehouse() {
-        Article product1 = new Tablet("Susa", "VK64", "Yep, it's a tablet", 10.5, 64, 80, 59.99,
-                GenerateUniqueId.generateUniqueID());
-        Warehouse.warehouse.add(product1);
-        Article product2 = new Tablet("SpyPhone", "69", "James Bond approves", 2, 2048, 6000, 10000,
-                GenerateUniqueId.generateUniqueID());
-        Warehouse.warehouse.add(product2);
-        Article product3 = new Smartphone("Big", "Chungus", "It's big", 20, 512, 400, 800,
-                GenerateUniqueId.generateUniqueID());
-        Warehouse.warehouse.add(product3);
-        Article product4 = new Smartphone("Susa", "Honor", "Good camera", 6, 128, 250, 500,
-                GenerateUniqueId.generateUniqueID());
-        Warehouse.warehouse.add(product4);
-        Article product5 = new Notebook("SpyPhone", "SlimJim", "We make notebooks too", 13, 4096, 8999, 24999,
-                GenerateUniqueId.generateUniqueID());
-        Warehouse.warehouse.add(product5);
-        Article product6 = new Notebook("Tencent", "YadMYqRiN", "It's cheap ;)", 14, 32, 0.05, 0.1,
-                GenerateUniqueId.generateUniqueID());
-        Warehouse.warehouse.add(product6);
-
-        Tablet tablet1 = new Tablet("Huawei", "xxx", "bello", 22, 150, 50, 160, GenerateUniqueId.generateUniqueID());
-        Warehouse.warehouse.add(tablet1);
-        Tablet tablet2 = new Tablet("Hp", "x222", "brutto", 10, 240, 80, 200, GenerateUniqueId.generateUniqueID());
-        Warehouse.warehouse.add(tablet2);
-        Smartphone smartphone1 = new Smartphone("Nokia", "3210", null, 11, 90, 150, 300,
-                GenerateUniqueId.generateUniqueID());
-        Warehouse.warehouse.add(smartphone1);
-        Smartphone smartphone2 = new Smartphone("Iphone", "12", "Funziona", 31, 120, 150, 400,
-                GenerateUniqueId.generateUniqueID());
-        Warehouse.warehouse.add(smartphone2);
-        Notebook notebook1 = new Notebook("Hp", "h300", null, 10, 222, 250, 600, GenerateUniqueId.generateUniqueID());
-        Warehouse.warehouse.add(notebook1);
-        Notebook notebook2 = new Notebook("Apple", "pro", "vecchio", 22, 400, 300, 800,
-                GenerateUniqueId.generateUniqueID());
-        Warehouse.warehouse.add(notebook2);
-    }
 }
