+++ carello/src/progetto/Warehouse.java
@@ -1,9 +1,6 @@
 package progetto;
 
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Objects;
-import java.util.Scanner;
+import java.util.*;
 
 public class Warehouse {
     static List<Article> warehouse = new ArrayList<>();
@@ -26,7 +23,7 @@ public class Warehouse {
 
         Scanner input = new Scanner(System.in);
         System.out.println("Enter the TYPE of the product");
-        String inputType = input.nextLine();
+        String inputType = input.nextLine().toUpperCase();
         Article.TypeOfArticle type = null;
         if (Objects.equals(inputType, "TABLET")) {
             type = Article.TypeOfArticle.TABLET;
@@ -74,7 +71,7 @@ public class Warehouse {
     public static void searchType() throws Exception {
         Scanner input = new Scanner(System.in);
         System.out.print("Please enter the TYPE of the product you wish to search for: ");
-        String userType = input.nextLine();
+        String userType = input.nextLine().toUpperCase();
         for (int i = 0; i < Warehouse.warehouse.size(); i++) {
             if (Objects.equals(userType, "TABLET")) {
                 if (Warehouse.warehouse.get(i).getType() == Article.TypeOfArticle.TABLET) {
@@ -100,10 +97,11 @@ public class Warehouse {
         warehouse.add(a);
     }
 
-    // Un metodo che permetta la ricerca per prezzo di acquisto
-    // Dovrà resitutire la lista di dispositivi frutto della ricerca
-    // o un errore nel caso in cui la ricerca non produca risultati.
-    public static List<Article> findBuyingPrice(int price) throws Exception {
+
+    public static List<Article> findBuyingPrice() throws Exception {
+        Scanner input = new Scanner(System.in);
+        System.out.println("Insert purchase price to find");
+        int price = input.nextInt();
         List<Article> listPrice = new ArrayList<Article>();
         for (Article a : warehouse) {
             if (a.getPriceOfBuying() == price) {
@@ -111,17 +109,19 @@ public class Warehouse {
             }
         }
         if (listPrice.isEmpty()) {
-            throw new Exception("Nessun prodotto a quel prezzo di acquisto in magazzino");
+            throw new Exception("There are no product with this purchase price in the warehouse");
         } else {
             return listPrice;
         }
     }
 
-    // Un metodo che permetta la ricerca in un determinato range di prezzo
-    // Dovrà resitutire la lista di dispositivi frutto della ricerca
-    // o un errore nel caso in cui la ricerca non produca risultati.
-    public static List<Article> findRangePrice(int priceMin, int priceMax) throws Exception {
+    public static List<Article> findRangePrice() throws Exception {
+        Scanner input = new Scanner(System.in);
         List<Article> listRange = new ArrayList<Article>();
+        System.out.println("Find product between this selling range price. \n Insert minimum price");
+        int priceMin = input.nextInt();
+        System.out.println("Insert maximum price");
+        int priceMax = input.nextInt();
         for (Article a : warehouse) {
             if (a.getPriceOfSelling() >= priceMin && a.getPriceOfSelling() <= priceMax) {
                 listRange.add(a);
@@ -134,12 +134,11 @@ public class Warehouse {
         }
     }
 
-    // Un metodo che calcoli la spesa media di acquisto per singolo dispositivo
-    // e che restituisca il valore medio.
-    // dovrà resitutire la lista di dispositivi frutto della ricerca
-    // o un errore nel caso in cui la ricerca non produca risultati.
-    public static double findAvgPrice(String type) throws Exception {
-        List<Article> listaVG = new ArrayList<Article>();
+    public static double findAvgPrice() throws Exception {
+        //TODO gestire se viene inserito un articolo inesistente
+        Scanner input = new Scanner(System.in);
+        System.out.println("Insert type of Article");
+        String type = input.next().toUpperCase();
         double avg = 0;
         int numArt = 0;
         for (Article art : warehouse) {
@@ -148,8 +147,15 @@ public class Warehouse {
                 numArt++;
             }
         }
-        return avg / numArt;
+        avg = avg / numArt;
+        if (avg != 0) {
+            return avg;
+        } else {
+            throw new Exception("This type of Article is not available in the warehouse");
+        }
+
     }
+
 
     public static List<Article> findByManifacturer() throws Exception {
 
@@ -205,5 +211,42 @@ public class Warehouse {
         } else {
             return listPriceOfSelling;
         }
+    }
+
+    public static void productsAlreadyInWarehouse() {
+        Article product1 = new Tablet("Susa", "VK64", "Yep, it's a tablet", 10.5, 64, 80, 59.99,
+                GenerateUniqueId.generateUniqueID());
+        Warehouse.warehouse.add(product1);
+        Article product2 = new Tablet("SpyPhone", "69", "James Bond approves", 2, 2048, 6000, 10000,
+                GenerateUniqueId.generateUniqueID());
+        Warehouse.warehouse.add(product2);
+        Article product3 = new Smartphone("Big", "Chungus", "It's big", 20, 512, 400, 800,
+                GenerateUniqueId.generateUniqueID());
+        Warehouse.warehouse.add(product3);
+        Article product4 = new Smartphone("Susa", "Honor", "Good camera", 6, 128, 250, 500,
+                GenerateUniqueId.generateUniqueID());
+        Warehouse.warehouse.add(product4);
+        Article product5 = new Notebook("SpyPhone", "SlimJim", "We make notebooks too", 13, 4096, 8999, 24999,
+                GenerateUniqueId.generateUniqueID());
+        Warehouse.warehouse.add(product5);
+        Article product6 = new Notebook("Tencent", "YadMYqRiN", "It's cheap ;)", 14, 32, 0.05, 0.1,
+                GenerateUniqueId.generateUniqueID());
+        Warehouse.warehouse.add(product6);
+
+        Tablet tablet1 = new Tablet("Huawei", "xxx", "bello", 22, 150, 50, 160, GenerateUniqueId.generateUniqueID());
+        Warehouse.warehouse.add(tablet1);
+        Tablet tablet2 = new Tablet("Hp", "x222", "brutto", 10, 240, 80, 200, GenerateUniqueId.generateUniqueID());
+        Warehouse.warehouse.add(tablet2);
+        Smartphone smartphone1 = new Smartphone("Nokia", "3210", null, 11, 90, 150, 300,
+                GenerateUniqueId.generateUniqueID());
+        Warehouse.warehouse.add(smartphone1);
+        Smartphone smartphone2 = new Smartphone("Iphone", "12", "Funziona", 31, 120, 150, 400,
+                GenerateUniqueId.generateUniqueID());
+        Warehouse.warehouse.add(smartphone2);
+        Notebook notebook1 = new Notebook("Hp", "h300", null, 10, 222, 250, 600, GenerateUniqueId.generateUniqueID());
+        Warehouse.warehouse.add(notebook1);
+        Notebook notebook2 = new Notebook("Apple", "pro", "vecchio", 22, 400, 300, 800,
+                GenerateUniqueId.generateUniqueID());
+        Warehouse.warehouse.add(notebook2);
     }
 }
